changer le style et ajouter la barre de recherche et ajouter l'arlert de confirmation

+++ view_data.php
@@ -141,8 +141,7 @@ try {
         </div>
     <form method="post" class="mb-3">
             <div class="form-group">
-                <label for="searchNumInscription">البحث برقم التسجيل:</label>
-                <input type="text" class="form-control" id="searchNumInscription" name="searchNumInscription" value="<?php echo $searchNumInscription; ?>">
+                <input type="text" class="form-control" id="searchNumInscription" name="searchNumInscription" value="<?php echo $searchNumInscription; ?>" placeholder="أدخل رقم التسجيل:">
             </div>
             <button type="submit" class="btn btn-primary" name="search">ابحث<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-search" viewBox="0 0 16 16">
               <path d="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001q.044.06.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1 1 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0"/>
@@ -194,7 +193,7 @@ try {
                            </svg>
                                 تحرير
                             </a>
-                            <a href="delete_data.php?NumInscription=<?php echo $row["NumInscription"]; ?>" class="btn btn-danger btn-sm">
+                            <a href="delete_data.php?NumInscription=<?php echo $row["NumInscription"]; ?>" class="btn btn-danger btn-sm" onclick="return confirmDelete();">
                             <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-person-fill-x" viewBox="0 0 16 16">
                                 <path d="M11 5a3 3 0 1 1-6 0 3 3 0 0 1 6 0m-9 8c0 1 1 1 1 1h5.256A4.5 4.5 0 0 1 8 12.5a4.5 4.5 0 0 1 1.544-3.393Q8.844 9.002 8 9c-5 0-6 3-6 4"/>
                                  <path d="M12.5 16a3.5 3.5 0 1 0 0-7 3.5 3.5 0 0 0 0 7m-.646-4.854.646.647.646-.647a.5.5 0 0 1 .708.708l-.647.646.647.646a.5.5 0 0 1-.708.708l-.646-.647-.646.647a.5.5 0 0 1-.708-.708l.647-.646-.647-.646a.5.5 0 0 1 .708-.708"/>
@@ -215,6 +214,11 @@ try {
             </tbody>
         </table>
     </div>
+    <script>
+    function confirmDelete() {
+        return confirm('هل أنت متأكد أنك تريد حذف هذا السجل؟');
+    }
+    </script>
 </body>
 
 </html>
